+++ src/main/java/com/adit/backend/domain/user/enums/Role.java
@@ -1,10 +1,11 @@
-package com.adit.backend.domain.user.entity;
+package com.adit.backend.domain.user.enums;
 
+import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
-@RequiredArgsConstructor
+@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
 public enum Role {
 	USER("ROLE_USER"),
 	ADMIN("ROLE_ADMIN");
